Fix bad url creation and don't ignore 401
In #64 it was found that `hass-cli info` would return empty result
when there was 401 auth errors AND for some reason setting HASSIO_TOKEN was not
enough to make use of http://hassio/homeassistant as base url.

401 was easy fix by simply ensuring we will throw
an exception if error occurs + added basic test for this.

Hassio issue was because urljoin was used for combing baseurl with api
path. Now moved to basic string concat so it will work as expected, i.e.

http://hassio/homeassistant + /api/discovery_info gives http://hassio/homeassistant/api/discovery_info

And not http://hassio/api/discovery_info which was the cause of the bugs.